massa_bootstrap::bindings

Module server

Structs

BootstrapServerBinder

Bootstrap server binder

ClientMessageLeader 🔒

The known-length component of a message to be received.

ErrorSendSlot 🔒

Releases a claimed ONGOING_ERROR_SENDS slot, panics included, so that a failing refusal can never leak the budget it took.

Constants

KNOWN_PREFIX_FROM_CLIENT_LEN 🔒

MAX_CONCURRENT_ERROR_SENDS 🔒

Maximum number of refusal threads running concurrently.

Statics

ONGOING_ERROR_SENDS 🔒

Number of refusal threads currently alive, compared against MAX_CONCURRENT_ERROR_SENDS.