Struct massa_bootstrap::bindings::server::BootstrapServerBinder

pub struct BootstrapServerBinder {
    max_consensus_block_ids: u64,
    thread_count: u8,
    max_datastore_key_length: u8,
    randomness_size_bytes: usize,
    local_keypair: KeyPair,
    duplex: Limiter<TcpStream>,
    prev_message: Option<Hash>,
    version_serializer: VersionSerializer,
    version_deserializer: VersionDeserializer,
    write_error_timeout: MassaTime,
}

Bootstrap server binder

Fields
max_consensus_block_ids: u64thread_count: u8max_datastore_key_length: u8randomness_size_bytes: usizelocal_keypair: KeyPairduplex: Limiter<TcpStream>prev_message: Option<Hash>version_serializer: VersionSerializerversion_deserializer: VersionDeserializerwrite_error_timeout: MassaTime
Implementations

impl BootstrapServerBinder

pub fn new( duplex: TcpStream, local_keypair: KeyPair, cfg: BootstrapSrvBindCfg, rw_limit: Option<u64> ) -> Self

Creates a new WriteBinder.

Argument

duplex: duplex stream.

local_keypair: local node user keypair

limit: limit max bytes per second (up and down)

pub fn handshake_timeout( &mut self, version: Version, duration: Option<Duration> ) -> Result<(), BootstrapError>

Performs a handshake. Should be called after connection MUST always be followed by a send of the BootstrapMessage::BootstrapTime

pub fn send_msg( &mut self, timeout: Duration, msg: BootstrapServerMessage ) -> Result<(), BootstrapError>

pub(crate) fn close_and_send_error<F>( self, msg: String, addr: SocketAddr, close_fn: F )where F: FnOnce() + Send + 'static,

Spawns a thread

blocks on the passed in runtime

uses passed in handle to send a message to the client

logs an error if the send times out

runs the passed in closure (typically a custom logging msg)

consumes the binding in the process

pub fn send_error_timeout( &mut self, error: String ) -> Result<(), BootstrapError>

pub fn send_timeout( &mut self, msg: BootstrapServerMessage, duration: Option<Duration> ) -> Result<(), BootstrapError>

Writes the next message.

pub fn next_timeout( &mut self, duration: Option<Duration> ) -> Result<BootstrapClientMessage, BootstrapError>

Read a message sent from the client (not signed).

fn decode_message_leader( &self, leader_buf: &[u8] ) -> Result<ClientMessageLeader, BootstrapError>

We are using this instead of of our library deserializer as the process is relatively straight forward and makes error-type management cleaner