massa_bootstrap

Module client

Structs

DefaultConnector

Initiates a connection with given timeout in milliseconds

Traits

BSConnector

Specifies a common interface that can be used by standard, or mockers

Functions

align_consensus_resume_state_before_ask 🔒

bootstrap_from_server 🔒

Gets the state from a bootstrap server (internal private function) needs to be CANCELLABLE

capped_consensus_cursor_ids 🔒

Ids for the reconnect ask: all finals if under the cap, otherwise the newest max_ids by slot.

capped_consensus_resume_step 🔒

check_restart_metadata 🔒

Checks the network restart metadata announced by a bootstrap server.

connect_to_server 🔒

filter_bootstrap_list 🔒

get_bootstrap_list_iter 🔒

get_state

Uses the cond-var pattern to handle sig-int cancellation. Make sure that the passed in interrupted shares its Arc with a sig-int handler setup.

restart_state_streaming 🔒

Drop everything streamed from the current server so that the next attempt streams the state again from scratch, instead of resuming on top of what this server sent.

send_client_message 🔒

shutdown_range 🔒

Computes the slot range of the last network shutdown, as startup does: from the slot right after the last one executed before the downtime, to the slot right before the restart.

stream_final_state_and_consensus 🔒

This function will send the starting point to receive a stream of the ledger and will receive and process each part until receive a BootstrapServerMessage::FinalStateFinished message from the server. next_bootstrap_message passed as parameter must be BootstrapClientMessage::AskFinalStatePart enum variant. next_bootstrap_message will be updated after receiving each part so that in case of connection lost we can restart from the last message we processed.

warn_user_about_versioning_updates 🔒