massa_bootstrap::client

Function restart_state_streaming

fn restart_state_streaming(
    client: &mut BootstrapClientBinder,
    next_bootstrap_message: &mut BootstrapClientMessage,
    global_bootstrap_state: &mut GlobalBootstrapState,
)

Drop everything streamed from the current server so that the next attempt streams the state again from scratch, instead of resuming on top of what this server sent.