massa_bootstrap::client

Trait BSConnector

pub trait BSConnector {
    // Required method
    fn connect_timeout(
        &self,
        addr: SocketAddr,
        duration: Option<MassaTime>,
    ) -> Result<TcpStream>;
}

Specifies a common interface that can be used by standard, or mockers

Required Methods

fn connect_timeout( &self, addr: SocketAddr, duration: Option<MassaTime>, ) -> Result<TcpStream>

The client attempts to connect to the given address. If a duration is provided, the attempt will be timed out after the given duration.