Enum massa_bootstrap::BootstrapServerMessage

pub enum BootstrapServerMessage {
    BootstrapTime {
        server_time: MassaTime,
        version: Version,
    },
    BootstrapPeers {
        peers: BootstrapPeers,
    },
    BootstrapPart {
        slot: Slot,
        state_part: StreamBatch<Slot>,
        versioning_part: StreamBatch<Slot>,
        consensus_part: BootstrapableGraph,
        consensus_outdated_ids: PreHashSet<BlockId>,
        last_start_period: Option<u64>,
        last_slot_before_downtime: Option<Option<Slot>>,
    },
    BootstrapFinished,
    SlotTooOld,
    BootstrapError {
        error: String,
    },
}

Messages used during bootstrap by server

Variants

BootstrapTime

Sync clocks

Fields

server_time: MassaTime

The current time on the bootstrap server.

version: Version

The version of the bootstrap server.

BootstrapPeers

Bootstrap peers

Fields

peers: BootstrapPeers

Server peers

BootstrapPart

Part of final state and consensus

Fields

slot: Slot

Slot the state changes are attached to

state_part: StreamBatch<Slot>

Part of the state in a serialized way

versioning_part: StreamBatch<Slot>

Part of the state (specific to versioning) in a serialized way

consensus_part: BootstrapableGraph

Part of the consensus graph

consensus_outdated_ids: PreHashSet<BlockId>

Outdated block ids in the current consensus graph bootstrap

last_start_period: Option<u64>

Last Start Period for network restart management

last_slot_before_downtime: Option<Option<Slot>>

Last Slot before downtime for network restart management

BootstrapFinished

Message sent when the final state and consensus bootstrap are finished

SlotTooOld

Slot sent to get state changes is too old

BootstrapError

Bootstrap error

Fields

error: String

Error message