Function massa_bootstrap::start_bootstrap_server

pub async fn start_bootstrap_server(
    consensus_controller: Box<dyn ConsensusController>,
    network_command_sender: NetworkCommandSender,
    final_state: Arc<RwLock<FinalState>>,
    bootstrap_config: BootstrapConfig,
    establisher: Establisher,
    keypair: KeyPair,
    version: Version
) -> Result<Option<BootstrapManager>, BootstrapError>

TODO merging the command senders into one channel structure may allow removing that allow

start a bootstrap server. Once your node will be ready, you may want other to bootstrap from you.