Crate massa_bootstrap

Re-exports

pub use establisher::Establisher;

Modules

establisher

settings

Structs

BootstrapManager

GlobalBootstrapState

a collection of the bootstrap state snapshots of all relevant modules

Functions

get_state

Gets the state from a bootstrap server needs to be CANCELLABLE

start_bootstrap_server