Crate massa_bootstrap

Copyright (c) 2022 MASSA LABS info@massa.net Bootstrap crate

At start up, if now is after genesis timestamp, the node will bootstrap from one of the provided bootstrap servers.

On server side, the server will query consensus for the graph and the ledger, execution for execution related data and network for the peer list.

Modules

bindings 🔒

client 🔒

error 🔒

listener 🔒

messages 🔒

server 🔒

start the bootstrapping system using start_bootstrap_server Once your node will be ready, you may want other to bootstrap from you.

settings 🔒

tools 🔒

white_black_list

white/black list

Structs

BootstrapClientMessageDeserializer

Deserializer for BootstrapClientMessage

BootstrapClientMessageSerializer

Serializer for BootstrapClientMessage

BootstrapConfig

Bootstrap configuration.

BootstrapManager

Abstraction layer over data produced by the listener, and transported over to the worker via a channel handle on the bootstrap server

BootstrapServerMessageDeserializer

Deserializer for BootstrapServerMessage

BootstrapServerMessageDeserializerArgs

Bootstrap Message der args

BootstrapServerMessageSerializer

Serializer for BootstrapServerMessage

BootstrapTcpListener

TODO: this should be crate-private. currently needed for models testing

DefaultConnector

Initiates a connection with given timeout in milliseconds

GlobalBootstrapState

a collection of the bootstrap state snapshots of all relevant modules

Enums

BootstrapClientMessage

Messages used during bootstrap by client

BootstrapError

Encapsulates the various failure contexts for the bootstrap process, both client and server side

BootstrapServerMessage

Messages used during bootstrap by server

IpType

Bootstrap IP protocol version setting.

Functions

get_state

Uses the cond-var pattern to handle sig-int cancellation. Make sure that the passed in interrupted shares its Arc with a sig-int handler setup.

start_bootstrap_server

See module level documentation for details