Crate massa_bootstrap

Copyright (c) 2022 MASSA LABS info@massa.net Bootstrap crate

At start up, if now is after genesis timestamp, the node will bootstrap from one of the provided bootstrap servers.

On server side, the server will query consensus for the graph and the ledger, execution for execution related data and network for the peer list.

Modules

types

Connection types

Structs

BootstrapClientMessageDeserializer

Deserializer for BootstrapClientMessage

BootstrapClientMessageSerializer

Serializer for BootstrapClientMessage

BootstrapConfig

Bootstrap configuration.

BootstrapManager

handle on the bootstrap server

BootstrapServerMessageDeserializer

Deserializer for BootstrapServerMessage

BootstrapServerMessageSerializer

Serializer for BootstrapServerMessage

GlobalBootstrapState

a collection of the bootstrap state snapshots of all relevant modules

Enums

BootstrapClientMessage

Messages used during bootstrap by client

BootstrapServerMessage

Messages used during bootstrap by server

IpType

Bootstrap IP protocol version setting.

Functions

get_state

Gets the state from a bootstrap server needs to be CANCELLABLE

start_bootstrap_server

TODO merging the command senders into one channel structure may allow removing that allow

Type Definitions

Establisher

connection establisher