Struct massa_bootstrap::settings::BootstrapConfig

pub struct BootstrapConfig {

    pub bootstrap_list: Vec<(SocketAddr, NodeId)>,
    pub bootstrap_protocol: IpType,
    pub bootstrap_whitelist_path: PathBuf,
    pub bootstrap_blacklist_path: PathBuf,
    pub listen_addr: Option<SocketAddr>,
    pub connect_timeout: MassaTime,
    pub bootstrap_timeout: MassaTime,
    pub read_timeout: MassaTime,
    pub write_timeout: MassaTime,
    pub read_error_timeout: MassaTime,
    pub write_error_timeout: MassaTime,
    pub retry_delay: MassaTime,
    pub max_ping: MassaTime,
    pub max_clock_delta: MassaTime,
    pub cache_duration: MassaTime,
    pub keep_ledger: bool,
    pub max_simultaneous_bootstraps: u32,
    pub per_ip_min_interval: MassaTime,
    pub ip_list_max_size: usize,
    pub max_bytes_read_write: u64,
    pub thread_count: u8,
    pub periods_per_cycle: u64,
    pub max_datastore_key_length: u8,
    pub randomness_size_bytes: usize,
    pub endorsement_count: u32,
    pub max_advertise_length: u32,
    pub max_listeners_per_peer: u32,
    pub max_bootstrap_blocks_length: u32,
    pub max_operations_per_block: u32,
    pub max_bootstrap_error_length: u64,
    pub max_new_elements: u64,
    pub max_datastore_entry_count: u64,
    pub max_datastore_value_length: u64,
    pub max_op_datastore_entry_count: u64,
    pub max_op_datastore_key_length: u8,
    pub max_op_datastore_value_length: u64,
    pub max_async_pool_changes: u64,
    pub max_async_pool_length: u64,
    pub max_async_message_data: u64,
    pub max_function_name_length: u16,
    pub max_parameters_size: u32,
    pub max_ledger_changes_count: u64,
    pub max_changes_slot_count: u64,
    pub max_rolls_length: u64,
    pub max_production_stats_length: u64,
    pub max_credits_length: u64,
    pub max_executed_ops_length: u64,
    pub max_ops_changes_length: u64,
    pub consensus_bootstrap_part_size: u64,
    pub max_consensus_block_ids: u64,
    pub mip_store_stats_block_considered: usize,
    pub max_denunciations_per_block_header: u32,
    pub max_denunciation_changes_length: u64,

}

Bootstrap configuration.

Fields
bootstrap_list: Vec<(SocketAddr, NodeId)>

Ip address of our bootstrap nodes and their public key.

bootstrap_protocol: IpType

IP version filter for bootstrap list, targeting IpType::IPv4, IpType::IPv6 or IpType::Both. Defaults to IpType::Both.

bootstrap_whitelist_path: PathBuf

Path to the bootstrap whitelist file. This whitelist define IPs that can bootstrap on your node.

bootstrap_blacklist_path: PathBuf

Path to the bootstrap blacklist file. This whitelist define IPs that will not be able to bootstrap on your node. This list is optional.

listen_addr: Option<SocketAddr>

Port to listen if we choose to allow other nodes to use us as bootstrap node.

connect_timeout: MassaTime

connection timeout

bootstrap_timeout: MassaTime

Time allocated to managing the bootstrapping process, i.e. providing the ledger and consensus

read_timeout: MassaTime

readout timeout

write_timeout: MassaTime

write timeout

read_error_timeout: MassaTime

readout error timeout

write_error_timeout: MassaTime

write error timeout

retry_delay: MassaTime

Time we wait before retrying a bootstrap

max_ping: MassaTime

Max ping delay.

max_clock_delta: MassaTime

Maximum allowed time between server and client clocks

cache_duration: MassaTime

Cache duration

keep_ledger: bool

Keep ledger or not if not bootstrap

max_simultaneous_bootstraps: u32

Max simultaneous bootstraps

per_ip_min_interval: MassaTime

Minimum interval between two bootstrap attempts from a given IP

ip_list_max_size: usize

Max size of the IP list

max_bytes_read_write: u64

Read-Write limitation for a connection in bytes per seconds

thread_count: u8

thread count

periods_per_cycle: u64

period per cycle

max_datastore_key_length: u8

max datastore key length

randomness_size_bytes: usize

randomness size bytes

endorsement_count: u32

endorsement count

max_advertise_length: u32

max advertise length

max_listeners_per_peer: u32

max listeners per peer

max_bootstrap_blocks_length: u32

max bootstrap blocks length

max_operations_per_block: u32

max operations per block

max_bootstrap_error_length: u64

max bootstrap error length

max_new_elements: u64

max bootstrap final state new_elements

max_datastore_entry_count: u64

max datastore entry count

max_datastore_value_length: u64

max datastore value length

max_op_datastore_entry_count: u64

max op datastore entry count

max_op_datastore_key_length: u8

max op datastore key length

max_op_datastore_value_length: u64

max op datastore value length

max_async_pool_changes: u64

max async pool changes

max_async_pool_length: u64

max async pool length

max_async_message_data: u64

max data async message

max_function_name_length: u16

max function name length

max_parameters_size: u32

max parameters size

max_ledger_changes_count: u64

max ledger changes

max_changes_slot_count: u64

max slot count in state changes

max_rolls_length: u64

max rolls in proof-of-stake and state changes

max_production_stats_length: u64

max production stats in proof-of-stake and state changes

max_credits_length: u64

max credits in proof-of-stake and state changes

max_executed_ops_length: u64

max executed ops

max_ops_changes_length: u64

max executed ops changes

consensus_bootstrap_part_size: u64

consensus bootstrap part size

max_consensus_block_ids: u64

max number of consensus block ids when sending a bootstrap cursor from the client

mip_store_stats_block_considered: usize

block count to check / process for versioning stats

max_denunciations_per_block_header: u32

max denunciations in block header

max_denunciation_changes_length: u64

max executed denunciations changes